Flytter HighlightedValues og Description ut i egne filer
