+++ src/web/screens/AffiliationViews/DeploymentView/DetailsView/VersionView/containers/VersionTable/VersionTable.tsx
@@ -10,6 +10,8 @@ import {
   IVersion,
 } from '../../../../../../../services/auroraApiClients/imageRepositoryClient/query';
 import { skeColor } from '@skatteetaten/frontend-components/utils';
+import HighlightedValues from './HighlightedValues';
+import Description from './Description';
 
 const columns = [
   {
@@ -122,30 +124,14 @@ export const VersionTable = ({
       ),
     }));
 
-  const Description = () => {
-    const { length } = versions;
-    const prefix = length > 0 ? `Viser ${length}` : 'Fant ingen';
-    const [er, e] = length === 1 ? ['', ''] : ['er', 'e'];
-
-    // prettier-ignore
-    // Prettier vil legge inn unødvendig mange linjeskift her. Denne trenger ikke å ta så mye plass.
-    switch (versionType) {
-      case ImageTagType.AURORA_VERSION: return <>{prefix} Aurora-versjon{er}.</>;
-      case ImageTagType.AURORA_SNAPSHOT_VERSION: return <>{prefix} unik{e} <i>snapshot</i>-versjon{er}.</>;
-      case ImageTagType.BUGFIX: return <>{prefix} <i>bugfix</i>-versjon{er}.</>;
-      case ImageTagType.LATEST: return <>{prefix} <i>latest</i>-versjon{er}.</>;
-      case ImageTagType.MAJOR: return <>{prefix} <i>major</i>-versjon{er}.</>;
-      case ImageTagType.MINOR: return <>{prefix} <i>minor</i>-versjon{er}.</>;
-      case ImageTagType.SNAPSHOT: return <>{prefix} <i>snapshot</i>-versjon{er}.</>;
-      case ImageTagType.COMMIT_HASH: return <>{prefix} <i>commit hash</i>-versjon{er}.</>;
-      default: return <>{prefix} resultat{er} av søk etter "{searchText}".</>;
-    }
-  };
-
   return (
     <TableWrapper>
       <SearchInfo>
-        <Description />
+        <Description
+          versionType={versionType}
+          numberOfVersions={versions.length}
+          searchText={searchText}
+        />
       </SearchInfo>
       {versions.length > 0 && (
         <Table
@@ -217,78 +203,4 @@ const SearchInfo = styled.div`
   background-color: ${skeColor.lightBlue};
   color: ${skeColor.blue100};
   padding: 12px;
-`;
-
-const HighlightedValues = (props: { searchText: string; text: string }) => {
-  const { searchText, text } = props;
-  const lcText = text.toLowerCase();
-
-  interface IIndices {
-    start: number;
-    end: number;
-  }
-
-  // Finner indekser for treff i søk:
-  const indicesList: IIndices[] = searchText
-    .split(' ')
-    .map((value) => value.trim().toLowerCase())
-    .filter((v) => v.length > 0)
-    .flatMap((v) => {
-      let startIndex = 0,
-        index = 0,
-        indices: IIndices[] = [];
-      while ((index = lcText.indexOf(v, startIndex)) > -1) {
-        indices.push({ start: index, end: index + v.length });
-        startIndex = index + 1;
-      }
-      return indices;
-    })
-    .sort((a, b) => a!.start - b!.start);
-
-  if (indicesList.length === 0) return null;
-
-  // Fletter sammen overlappende treff:
-  let mergedList: IIndices[] = [],
-    current: IIndices = indicesList[0];
-  for (let i = 0; i < indicesList.length; i++) {
-    if (i === indicesList.length - 1) {
-      mergedList.push(current);
-    } else {
-      const next = indicesList[i + 1];
-      if (next.start < current.end) {
-        if (next.end > current.end) current.end = next.end;
-      } else {
-        mergedList.push(current);
-        current = next;
-      }
-    }
-  }
-
-  // Formaterer resultat:
-  const startOfString =
-    mergedList[0].start > 0 ? [<>{text.slice(0, mergedList[0].start)}</>] : [];
-  return (
-    <>
-      {startOfString.concat(
-        mergedList.flatMap((indices, i) => {
-          const matchingString = (
-            <Highlight key={`h${i}`}>
-              {text.slice(indices.start, indices.end)}
-            </Highlight>
-          );
-          const afterMatchingString = text.slice(
-            indices.end,
-            mergedList[i + 1]?.start
-          );
-          return afterMatchingString.length === 0
-            ? [matchingString]
-            : [matchingString, <>{afterMatchingString}</>];
-        })
-      )}
-    </>
-  );
-};
-
-const Highlight = styled.span`
-  background-color: ${skeColor.lightBrown};
 `;
